+++ src/architecture/Main.java
@@ -1,4 +1,4 @@
-package sample;
+package architecture;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
