+++ src/architecture/databases/HibConfig.java
@@ -1,11 +1,11 @@
-package sample.databases;
+package architecture.databases;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.cfg.Environment;
 import org.hibernate.service.ServiceRegistry;
-import sample.model.Contact;
+import architecture.model.Contact;
 
 import java.util.Properties;
