--- src/sample/controller/ContactController.java
@@ -1,76 +1,0 @@
-package sample.controller;
-
-import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
-import javafx.scene.control.*;
-import sample.databases.ContactDAO;
-import sample.model.Contact;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ContactController {
-    private List<Contact> contacts = new ArrayList<>();
-
-    @FXML
-    private MenuBar menuBar;
-
-    @FXML
-    private Menu file;
-
-    @FXML
-    private MenuItem close;
-
-    @FXML
-    private Menu help;
-
-    @FXML
-    private Button addContact;
-
-    @FXML
-    private TextField name;
-
-    @FXML
-    private TextField city;
-
-    @FXML
-    private TextField phone;
-
-    @FXML
-    private ListView<String> contactsList;
-
-    @FXML
-    private ContextMenu contextManu;
-
-    @FXML
-    private MenuItem menuItem;
-
-    @FXML
-    private Button showContacts;
-
-    @FXML
-    private void initialize() {
-        System.out.println("ContactController initialized!");
-    }
-
-    @FXML
-    void onBtnAddContactClick(ActionEvent event) {
-        System.out.println("Adding contact: " + name.getText() + city.getText() + phone.getText());
-
-        Contact contact = new Contact(name.getText(), city.getText(), Integer.parseInt(phone.getText()));
-        ContactDAO contactDAO = new ContactDAO();
-        contactDAO.createContact(contact);
-
-        contacts.add(contact);
-    }
-
-    @FXML
-    void onBtnShowContactsClick(ActionEvent event) {
-        System.out.println("Show Contacts Click");
-        ContactDAO contactDAO = new ContactDAO();
-        contacts = contactDAO.getContacts();
-        for (Contact con : contacts) {
-            contactsList.getItems().add(con.toString());
-        }
-    }
-}
